refactor(clean): import PayPalScriptPvider from index in stories
