+++ src/stories/PayPalButtons.stories.tsx
@@ -11,9 +11,8 @@ import type {
     OnApproveActions,
 } from "@paypal/paypal-js/types/components/buttons";
 
-import { FUNDING, PayPalButtons } from "../index";
+import { PayPalScriptProvider, PayPalButtons, FUNDING } from "../index";
 import { getOptionsFromQueryString, generateRandomString } from "./utils";
-import { PayPalScriptProvider } from "../components/PayPalScriptProvider";
 
 const scriptProviderOptions: PayPalScriptOptions = {
     "client-id": "test",
